fix: 修复能量/容量 int 截断与机械成本越界；新增机械三机并全量重制材质建模
- Issue1: 赤能/生命能容量经升级倍率后超 Integer.MAX_VALUE，SimpleContainerData 仅支持 int 会截断，
  改用 LongDataSlots 高低 32 位拆分同步（发生器矩阵、矿机控制器、药水台、词条重铸器等），
  配套 AkaishiEnergyStorage 支持 volatile long 上限与动态扩容。
- Issue2: MechanicalMachineCosts 越界时原返回 1L 会压低费用/工时导致近乎免费加工，
  改为回退内置默认数组末元素。
- 新增机械三机：模板制造厂/加工制作厂/组装加工台（方块、BlockEntity、菜单、Screen、成本与部件体系）。
- 修正菜单兜底占位槽数：装备打造器 3→12、加工制作厂 6→10，避免客户端读 data 越界。
- 全量重制材质与建模（方块/物品/装甲/流体/管道/GUI）并同步双语 lang、JEI、手册与创造栏。

+++ common/src/main/java/com/example/akaishi/menu/AkaishiEnergyAggregatorScreen.java
@@ -41,8 +41,8 @@ public class AkaishiEnergyAggregatorScreen extends AbstractContainerScreen<Akais
 
         // 赤能源条（红色）
         GuiWidgets.track(gui, x + BAR_X, y + BAR_Y, BAR_W, BAR_H);
-        long max = Math.max(1, menu.getMaxEnergy());
-        int energyWidth = (int) (BAR_W * Math.max(0, Math.min(menu.getEnergy(), menu.getMaxEnergy())) / max);
+        long max = Math.max(1L, menu.getMaxEnergy());
+        int energyWidth = (int) (BAR_W * Math.max(0L, Math.min(menu.getEnergy(), menu.getMaxEnergy())) / max);
         if (energyWidth > 0) {
             gui.fill(x + BAR_X, y + BAR_Y, x + BAR_X + energyWidth, y + BAR_Y + BAR_H, 0xFFE03030);
         }
@@ -64,7 +64,7 @@ public class AkaishiEnergyAggregatorScreen extends AbstractContainerScreen<Akais
                         EnergyFormat.format(menu.getEnergy()), EnergyFormat.format(menu.getMaxEnergy())),
                 20, 50, 0xE0E0E0, false);
         // 可合成次数文本（按当前配方消耗计算，母岩升级/赤石锭聚合通用）
-        int cost = Math.max(1, menu.getCurrentCost());
+        long cost = Math.max(1L, menu.getCurrentCost());
         gui.drawString(this.font,
                 Component.translatable("gui.akaishi.craft_times",
                         menu.getEnergy() / cost),
